- reworked font loader to make adding multi-lump fonts easier.
A multi-lump font can be created by putting all characters into a subdirectory of fonts/ with the intended name. Each character needs to be named by its character index as hex number.
So far this is only active for the predefined small fonts

+++ src/scripting/vmthunks.cpp
@@ -257,7 +257,7 @@ DEFINE_ACTION_FUNCTION_NATIVE(FStringStruct, ToInt, StringToInt)
 {
 	PARAM_SELF_STRUCT_PROLOGUE(FString);
 	PARAM_INT(base);
-	ACTION_RETURN_INT(self->ToLong(base));
+	ACTION_RETURN_INT((int)self->ToLong(base));
 }
 
 static double StringToDbl(FString *self)
